+++ src/components/Card/index.tsx
@@ -2,17 +2,18 @@ import React, { memo } from 'react';
 import Style from './style';
 
 type CardProps = {
-  readonly active: boolean;
-  readonly index: number;
-  readonly handleCardMouseDown: (index: number) => void;
+  active: boolean;
+  index: number;
+  handleCardMouseDown: (index: number) => void;
+  title: string;
 };
 
-const Card = ({ active, index, handleCardMouseDown }: CardProps): JSX.Element =>
+const Card = ({ active, index, handleCardMouseDown, title }: Readonly<CardProps>): JSX.Element =>
   <Style
     className={`card${active ? ' active' : ''}`}
     onClick={(event: React.MouseEvent<HTMLInputElement>): void => event.stopPropagation()}
     onMouseDown={(): void => handleCardMouseDown(index)}>
-    Hello
+    {title}
   </Style>;
 
 export default memo(Card);